1.0.2

+++ db/chats.py
@@ -13,7 +13,7 @@ ChatTable = sa.Table(
     'chats',
     METADATA,
     sa.Column('id', sa.Integer, primary_key=True, autoincrement=True),
-    sa.Column('chat_id', sa.Integer),
+    sa.Column('chat_id', sa.BigInteger),
 )
